Time: 133 ms (19.22%), Space: 32.3 MB (76.96%) - LeetHub

+++ 1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts/1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts.cpp
@@ -9,7 +9,7 @@ public:
         sort(horizontalCuts.begin(),horizontalCuts.end());
         sort(verticalCuts.begin(),verticalCuts.end());
         
-        long l=0,b=0;
+        long long l=0,b=0;
         for(int i=1;i<horizontalCuts.size();i++)
         {
             if(l<horizontalCuts[i]-horizontalCuts[i-1])
